requirements.txt
python-decouple
telethon
requests
Pillow>=4.2.0
telegraph
